Improve code for fake location ui
* Separate ui code for fake location manager to keep ui and logic
separate in code
* Improve comments in code to explain code better
